Windows support and version bump (0.4.1) TensorFlow support WIP. This should addresses https://github.com/BachiLi/redner/issues/7 and https://github.com/BachiLi/redner/issues/93

+++ src/buffer.h
@@ -93,8 +93,8 @@ public:
         }
     }
 
-    int size() const { return count; }
-    int bytes() const { return count * sizeof(T); }
+    int size() const { return (int)count; }
+    int bytes() const { return (int)count * sizeof(T); }
 
     T* begin() {
         return data;
